Created a single-step build process
The makefile calls the script to copy/update the Serpent files, then builds the library in the proper sequence
When this is implemented as a module in MOOSE then the global module makefile will need to have a check for SERPENT_HOME in order to trigger the build of this module

+++ scripts/update_serpent.sh
@@ -2,30 +2,37 @@
 # This script copies the files from ${SERPENT_HOME} to ./serpent and
 # prepares them to be compile compatibly for use within MOOSE
 
+# Acquire/define the directory for the local copy
 CURRENT_DIR=$(pwd)
-FOLDER=$(basename ${CURRENT_DIR})
-if [ "${FOLDER}" == "scripts" ]; then
-  cd ..
+COPY_TO_DIR=${1:-${CURRENT_DIR}/serpent}
+
+# Ensure our two directories are different
+if [[ ${COPY_TO_DIR} -ef ${SERPENT_HOME} ]]; then
+  printf "Cannot copy when the source and destination directories are the same!\n"
+  exit 1
 fi
-mkdir -p ./serpent
 
-# cd into serpent
-cd serpent
-
-printf "Copying source files from %s to %s... " ${SERPENT_HOME} $(pwd)
-cp ${SERPENT_HOME}/{*.h,*.c} .
-printf "done! \n"
+# Ensure the destination directory exists and then copy
+mkdir -p ${COPY_TO_DIR}
+cd ${COPY_TO_DIR}
+printf "Copying source files from %s to %s... \n" ${SERPENT_HOME} ${COPY_TO_DIR}
+cp -p ${SERPENT_HOME}/{*.h,*.c} .
 
 # Remove invalid beta files
 rm deffcordata.c
 
-printf "Updating .h files with include guard... "
+printf "Patching .h files... \n"
 for file in *.h; do
-  sed -i.old '1s;^;#pragma once\n\n;' ${file}
-done
-printf "done!\n"
+  # Only patch files that are suspect of being unmodified
+  if [[ ! ${file} -nt ${SERPENT_HOME}/${file} ]]; then
+    if [[ "${file}" == "header.h" ]]; then
+      # Put header.h in an extern "C" block
+      sed -i.old '1s;^;#ifdef __cplusplus\nextern "C"{\n#endif\n\n;' header.h
+      printf "#ifdef __cplusplus\n}\n#endif\n" >> header.h
+    fi
 
-printf 'Placing extern "C"{} block on "header.h"... '
-  sed -i.old '1s;^;#ifdef __cplusplus\nextern "C"{\n#endif\n\n;' header.h
-  printf "#ifdef __cplusplus\n}\n#endif\n" >> header.h
-printf "done!\n"
+    # Place an include guard on all header files
+    sed -i.old '1s;^;#pragma once\n\n;' ${file}
+    rm ${file}.old
+  fi
+done
